pyproject.toml
[tool.poetry]
name = "pycardano"
version = "0.1.0"
description = "A python implementation of Cardano client"
authors = ["Jerry <jerry@staking247.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
PyNaCl = "^1.4.0"
cbor2 = "^5.4.2"
typeguard = "^2.13.3"
blockfrost-python = "^0.4.2"

[tool.poetry.dev-dependencies]
Sphinx = "^4.3.2"
sphinx-rtd-theme = "^1.0.0"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.1.0"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--doctest-modules --ignore=examples"
minversion = "6.0"

[tool.isort]
profile = "black"
